Organize ATS export helpers by area
Move the misplaced ServiceCollectionExports helpers into eventing, health check, and Aspire store export files. Split the matching tests so they no longer reference the removed ServiceCollectionExports type.
