+++ tests/Aspire.Hosting.Tests/AtsEventingExportsTests.cs
@@ -3,23 +3,19 @@
 
 using System.Linq.Expressions;
 using System.Reflection;
-using Aspire.Hosting.Ats;
 using Aspire.Hosting.Lifecycle;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
-using AtsHealthCheckResult = Aspire.Hosting.Ats.HealthCheckResult;
 
 namespace Aspire.Hosting.Tests;
 
 [Trait("Partition", "5")]
-public class AtsServiceCollectionExportsTests
+public class AtsEventingExportsTests
 {
     [Fact]
     public void TryAddEventingSubscriber_AllowsDistinctCallbacks()
     {
         var builder = DistributedApplication.CreateBuilder([]);
         var method = typeof(DistributedApplication).Assembly
-            .GetType("Aspire.Hosting.Ats.ServiceCollectionExports", throwOnError: true)!
+            .GetType("Aspire.Hosting.Ats.EventingExports", throwOnError: true)!
             .GetMethod("TryAddEventingSubscriber", BindingFlags.Public | BindingFlags.Static)!;
         var callbackSubscriberType = method.DeclaringType!.GetNestedType("CallbackEventingSubscriber", BindingFlags.NonPublic)!;
 
@@ -37,28 +33,6 @@ public class AtsServiceCollectionExportsTests
             .ToList();
 
         Assert.Equal(2, subscribers.Count);
-    }
-
-    [Fact]
-    public async Task AddHealthCheck_RegistersCallback()
-    {
-        var builder = DistributedApplication.CreateBuilder([]);
-
-        builder.AddHealthCheck("custom_check", () => Task.FromResult(new AtsHealthCheckResult
-        {
-            Status = HealthStatus.Degraded,
-            Description = "custom description"
-        }));
-
-        using var serviceProvider = builder.Services.BuildServiceProvider();
-        var healthCheckService = serviceProvider.GetRequiredService<HealthCheckService>();
-
-        var report = await healthCheckService.CheckHealthAsync(registration => registration.Name == "custom_check");
-        var entry = Assert.Single(report.Entries);
-
-        Assert.Equal("custom_check", entry.Key);
-        Assert.Equal(HealthStatus.Degraded, entry.Value.Status);
-        Assert.Equal("custom description", entry.Value.Description);
     }
 
     private static Delegate CreateCallback(Type delegateType)
